# .migration-backup/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Enterprise Recruitment Automation Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /applications:
    get:
      operationId: listApplications
      tags: [applications]
      summary: List applications
      parameters:
        - name: jobId
          in: query
          schema:
            type: integer
        - name: candidateId
          in: query
          schema:
            type: integer
        - name: stage
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of applications
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicationList"
    post:
      operationId: createApplication
      tags: [applications]
      summary: Submit an application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationInput"
      responses:
        "201":
          description: Created application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
  /applications/{id}:
    get:
      operationId: getApplication
      tags: [applications]
      summary: Get application by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
    patch:
      operationId: updateApplication
      tags: [applications]
      summary: Update an application
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationUpdate"
      responses:
        "200":
          description: Updated application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
  /applications/{id}/move-stage:
    post:
      operationId: moveApplicationStage
      tags: [applications]
      summary: Move application to a new pipeline stage
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StageMove"
      responses:
        "200":
          description: Updated application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"

  # ─── INTERVIEWS ───────────────────────────────────────────────────────────────
...
  /audit-logs:
    get:
      operationId: listAuditLogs
      tags: [audit]
      summary: List audit log entries
      parameters:
        - name: entityType
          in: query
          schema:
            type: string
        - name: entityId
          in: query
          schema:
            type: integer
        - name: userId
          in: query
          schema:
            type: integer
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Audit log entries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditLogList"

  # ─── NOTIFICATIONS ────────────────────────────────────────────────────────────
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List notifications for current user
      parameters:
        - name: unreadOnly
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /notifications/{id}/read:
    post:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated notification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
  /admin/departments:
    get:
      operationId: listDepartments
      tags: [admin]
      summary: List departments
      responses:
        "200":
          description: Departments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Department"
    post:
      operationId: createDepartment
      tags: [admin]
      summary: Create a department
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepartmentInput"
      responses:
        "201":
          description: Created department
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
...
  /admin/users:
    get:
      operationId: listUsers
      tags: [admin]
      summary: List system users
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [admin]
      summary: Create a system user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /admin/users/{id}:
    patch:
      operationId: updateUser
      tags: [admin]
      summary: Update a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

# ─── COMPONENTS ────────────────────────────────────────────────────────────────
components:
  schemas:
...
    User:
      type: object
      required: [id, name, email, role]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [hr_admin, recruiter, hiring_manager, finance_approver, department_head, executive_viewer, candidate]
        department:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
    UserInput:
      type: object
      required: [name, email, role, password]
      properties:
        name:
          type: string
        email:
          type: string
        role:
          type: string
        department:
          type: ["string", "null"]
        password:
          type: string
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
        department:
          type: ["string", "null"]
        isActive:
          type: boolean

    # DASHBOARD
...
    Job:
      type: object
      required: [id, title, department, type, status, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        department:
          type: string
        description:
          type: ["string", "null"]
        competencies:
          type: array
          items:
            type: string
        type:
          type: string
          enum: [internal, external, both]
        status:
          type: string
          enum: [draft, active, closed, expired]
        source:
          type: ["string", "null"]
        expiryDate:
          type: ["string", "null"]
        requisitionId:
          type: ["integer", "null"]
        applicationCount:
          type: integer
        createdAt:
          type: string
...
    Candidate:
      type: object
      required: [id, firstName, lastName, email, source, createdAt]
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        currentRole:
          type: ["string", "null"]
        currentCompany:
          type: ["string", "null"]
        yearsOfExperience:
          type: ["number", "null"]
        skills:
          type: array
          items:
            type: string
        source:
          type: string
        resumeUrl:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        aiMatchScore:
          type: ["number", "null"]
        createdAt:
          type: string
...
    Application:
      type: object
      required: [id, candidateId, jobId, stage, appliedAt]
      properties:
        id:
          type: integer
        candidateId:
          type: integer
        jobId:
          type: integer
        candidate:
          $ref: "#/components/schemas/Candidate"
        job:
          $ref: "#/components/schemas/Job"
        stage:
          type: string
          enum: [applied, under_review, longlisted, shortlisted, interview_scheduled, interviewed, offer_pending, offered, hired, rejected]
        notes:
          type: ["string", "null"]
        rejectionReason:
          type: ["string", "null"]
        appliedAt:
          type: string
        updatedAt:
          type: string
    ApplicationInput:
      type: object
      required: [candidateId, jobId]
      properties:
        candidateId:
          type: integer
        jobId:
          type: integer
        notes:
          type: ["string", "null"]
    ApplicationUpdate:
      type: object
      properties:
        stage:
          type: string
        notes:
          type: ["string", "null"]
        rejectionReason:
          type: ["string", "null"]
    ApplicationList:
      type: object
      required: [data, total]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Application"
        total:
          type: integer
    StageMove:
      type: object
      required: [stage]
      properties:
        stage:
          type: string
        notes:
          type: ["string", "null"]

    # INTERVIEWS
...
    AuditLog:
      type: object
      required: [id, action, entityType, userId, userName, timestamp]
      properties:
        id:
          type: integer
        action:
          type: string
        entityType:
          type: string
        entityId:
          type: ["integer", "null"]
        userId:
          type: integer
        userName:
          type: string
        details:
          type: ["string", "null"]
        ipAddress:
          type: ["string", "null"]
        timestamp:
          type: string
    AuditLogList:
      type: object
      required: [data, total]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/AuditLog"
        total:
          type: integer

    # NOTIFICATIONS
    Notification:
      type: object
      required: [id, type, message, read, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [approval_required, interview_scheduled, offer_approved, sla_breach, onboarding_task, general]
        message:
          type: string
        entityType:
          type: ["string", "null"]
        entityId:
          type: ["integer", "null"]
        read:
          type: boolean
        createdAt:
          type: string

    # APPROVALS
...
    Department:
      type: object
      required: [id, name, headCount]
      properties:
        id:
          type: integer
        name:
          type: string
        headCount:
          type: integer
        headOfDepartment:
          type: ["string", "null"]
    DepartmentInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        headOfDepartment:
          type: ["string", "null"]
...