# .migration-backup/lib/api-zod/src/generated/types/auditLog.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Enterprise Recruitment Automation Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface AuditLog {
  id: number;
  action: string;
  entityType: string;
  /** @nullable */
  entityId?: number | null;
  userId: number;
  userName: string;
  /** @nullable */
  details?: string | null;
  /** @nullable */
  ipAddress?: string | null;
  timestamp: string;
}